jmwalletd/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jmwalletd"
version = "0.17.0"
description = "JoinMarket wallet daemon - JAM-compatible HTTP/WebSocket API"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.11"
dependencies = [
    "jmcore",
    "jmwallet",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.5.0",
    "pyjwt>=2.8.0",
    "loguru",
    "typer>=0.9.0",
    "httpx>=0.27.0",
    "websockets>=14.0",
    "aiohttp>=3.9.0",
    "coincurve>=20.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.scripts]
jmwalletd = "jmwalletd.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/jmwalletd/deps.py" = ["B008"]
"src/jmwalletd/routers/*.py" = ["B008"]
"tests/*.py" = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["jmwalletd", "jmcore", "jmwallet", "maker", "taker"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
